fix: Mark the gossip keys from the message as verified, not the ones from the db
Previously, `set_verified()` loaded the gossip key from the database and
then set it as verified. Now, if, for some reason, the database doesn't
contain the key that was just gossiped with the incoming message,
then we'll mark the wrong key as verified ("gossip key injection").

Instead, remember the `SignedPublicKey`s in the `MimeMessage` and
directly mark those as verified.