api: remove unused dc_accounts_all_work_done()
it was used by iOS to know when a background fetch was complete;
meanwhile the superiour dc_accounts_background_fetch() is used for that.

there is still the corresponding context function dc_all_work_done(),
this not used by any UI as well, however, it is in use by a python tests.