fix: Set Config::NotifyAboutWrongPw before saving configuration (#5896)
Let's not complicate the logic and always set `Config::NotifyAboutWrongPw` before saving
configuration, better if a wrong password notification is shown once more than not shown at all. It
shouldn't be a big problem because reconfiguration is a manual action and isn't done frequently.

Also for the same reason reset `Config::NotifyAboutWrongPw` only after a successful addition of the
appropriate device message.