feat: Don't put displayname into From/To/Sender if it equals to address
If a displayname equals to the address, adding it looks excessive. Moreover, it's not useful for
Delta Chat receiving the message because `sanitize_name_and_addr()` removes such a displayname
anyway. Also now at least DC Android requires specifying profile name, so there should be a fallback
for users having meaningful addresses to keep the old behaviour when Core generates `From` w/o the
profile name, and this question has already appeared on the forum, so maybe documenting it in FAQ
makes sense.