More context for the "Cannot establish guaranteed..." info message
The "Cannot establish guaranteed end-to-end encryption with ..." info
message can have lots of causes, and it happened twice to us now that it
took us some time to figure out which one it is.

So, include some more detail in the info message by simply adding the
non-translated error message in parantheses.

If we want to put in some more effort for nicer error messages, we
could:
- Introduce one new translated string
  "Cannot establish guaranteed end-to-end encryption with …. Cause: %2$s"
  or similar (and remove the old stock string)
- Introduce new translated strings for all the possible errors