fix: start progress bar for outgoing backup transfer early
Before this change progress bar only starts
when database is already transferred.
Database is usually the largest file
in the whole transfer, so the transfer appears
to be stuck for the sender.

`export_backup_stream` expects that progress 10
has already been emitted, but this was only the case
for export to the file.